+++ src/krisi/explore/analysis.py
