+++ src/krisi/report/interactive.py
